Version 2.0 - Name your file, and works with Grid OR list views
Added a textbox to the right of the button so you can change the name of the collection.  The default name will be the name of the collection.
Textbox added: Line 26
Textbox styled: Line 27
Default collection name set: Line 21

Previous versions also only worked if the collection was set to grid instead of list because I was focusing on the wrong element.  This has been corrected now as well (line 43)